// src/app/api/contribuyente/models/domicilio-relacionado-buscar-response.ts
/* tslint:disable */
/* eslint-disable */
export interface DomicilioRelacionadoBuscarResponse {
  activo?: number;
  contribuyenteNumero?: number;
  domRelacionadoNumero?: number;
  domicilioDsc?: string;
  estadoDsc?: string;
  index?: number;
  municipalidadId?: number;
  relContribuyenteNumero?: number;
  tipoPredioDsc?: string;
}