// src/app/api/contribuyente/models/domicilio-relacionado-response.ts
/* tslint:disable */
/* eslint-disable */
export interface DomicilioRelacionadoResponse {
  activo?: number;
  contribuyenteNumero?: number;
  departamentoId?: number;
  descripcionDomicilio?: string;
  descripcionInterior?: string;
  distritoId?: number;
  domRelacionadoNumero?: number;
  edificacionId?: number;
  estructurado?: number;
  fuenteInfoId?: number;
  ingreso?: string;
  kilometro?: string;
  latitud?: number;
  letra1?: string;
  letra2?: string;
  longitud?: number;
  lote?: string;
  manzana?: string;
  municipalidadId?: number;
  numero1?: number;
  numero2?: number;
  piso?: string;
  provinciaId?: number;
  referencia?: string;
  relContribuyenteNumero?: number;
  subLote?: string;
  subZonaUrbanaId?: number;
  tipoInteriorId?: number;
  tipoPredioId?: number;
  viaDepartamentoId?: number;
  viaDistritoId?: number;
  viaId?: number;
  viaProvinciaId?: number;
  zonaUrbanaId?: number;
}